requirements.txt
flask
flask-cors
sentence-transformers
faiss-cpu
google-generativeai
numpy